backend/app/requirements.txt
fastapi
uvicorn[standard]
playwright
pydantic
email-validator
requests
python-multipart
opencv-python-headless
numpy
pillow
python-dotenv
pytesseract
sqlalchemy
psycopg2-binary
alembic